+++ tests/source/meen_test/test_Machine.py
@@ -8,14 +8,14 @@ from meenPy import __version__
 from meenPy import ErrorCode
 from meenPy import Make8080Machine
 
-# import Python controller modules (a port of the c++ modules below)
-# always use the c++ memory controller module for performance reasons, the python module is available strictly for demonstration purposes
+# Import Python controller modules (a port of the c++ modules are available below)
+# Always use the c++ memory controller module for performance reasons, the python module is available strictly for demonstration purposes
 #from MemoryController import MemoryController
 from TestIoController import TestIoController
 from CpmIoController import CpmIoController
 
-# These c++ controller module versions are also available (CpmIoController currently not working)
-# always use the c++ memory controller for performance reasons
+# These c++ controller module versions are also available
+# Always use the c++ memory controller for performance reasons
 from TestControllersPy import MemoryController
 #from TestControllersPy import CpmIoController
 #from TestControllersPy import TestIoController
@@ -24,11 +24,11 @@ class MachineTest(unittest.TestCase):
     def setUp(self):
         self.programsDir = MachineTestDeps.programsDir
         self.memoryController = MemoryController()
-        self.cpmIoController = CpmIoController(self.memoryController)
+        self.cpmIoController = CpmIoController()
         self.testIoController = TestIoController()
         self.machine = Make8080Machine()
-        self.machine.SetIoController(self.testIoController)
-        self.machine.SetMemoryController(self.memoryController)
+        self.machine.AttachIoController(self.testIoController)
+        self.machine.AttachMemoryController(self.memoryController)
         err = self.machine.SetOptions(r'{"isrFreq":0.02}')
         self.assertEqual(err, ErrorCode.NoError)
         err = self.memoryController.Load(self.programsDir + 'exitTest.bin', 0x0000)
@@ -40,11 +40,11 @@ class MachineTest(unittest.TestCase):
         self.assertTrue(re.match(r'^(?P<major>0|[1-9]\d*)\.(?P<minor>0|[1-9]\d*)\.(?P<patch>0|[1-9]\d*)(?:-(?P<prerelease>(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+(?P<buildmetadata>[0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$', __version__))
 
     def test_SetNoIoController(self):
-        err = self.machine.SetIoController(None)
+        err = self.machine.AttachIoController(None)
         self.assertEqual(err, ErrorCode.InvalidArgument)
 
     def test_SetNoMemoryController(self):
-        err = self.machine.SetMemoryController(None)
+        err = self.machine.AttachMemoryController(None)
         self.assertEqual(err, ErrorCode.InvalidArgument)
 
     def test_NegativeISRFrequency(self):
@@ -59,15 +59,15 @@ class MachineTest(unittest.TestCase):
         self.assertEqual(err, ErrorCode.NoError)
         err = self.memoryController.Load(self.programsDir + 'nopEnd.bin', 0xC353)
         self.assertEqual(err, ErrorCode.NoError)
-        self.machine.Run(0)
-
-        # self.machine.Run(0x100)
+        
+        err = self.machine.Run(0x0000)
+        self.assertEqual(err, ErrorCode.NoError)
 
         err = self.machine.SetOptions(r'{"isrFreq":1}')
         self.assertEqual(err, ErrorCode.Busy)
-        err = self.machine.SetMemoryController(self.memoryController)
+        err = self.machine.AttachMemoryController(self.memoryController)
         self.assertEqual(err, ErrorCode.Busy)
-        err = self.machine.SetIoController(self.testIoController)
+        err = self.machine.AttachIoController(self.testIoController)
         self.assertEqual(err, ErrorCode.Busy)
         err = self.machine.OnSave(lambda: ErrorCode.NoError)
         self.assertIn(err, [ErrorCode.Busy, ErrorCode.NotImplemented])
@@ -78,9 +78,9 @@ class MachineTest(unittest.TestCase):
 
         err = self.machine.SetOptions(r'{"isrFreq":1}')
         self.assertEqual(err, ErrorCode.NoError)
-        err = self.machine.SetMemoryController(self.memoryController)
+        err = self.machine.AttachMemoryController(self.memoryController)
         self.assertEqual(err, ErrorCode.NoError)
-        err = self.machine.SetIoController(self.testIoController)
+        err = self.machine.AttachIoController(self.testIoController)
         self.assertEqual(err, ErrorCode.NoError)
         err = self.machine.OnSave(lambda: ErrorCode.NoError)
         self.assertIn(err, [ErrorCode.NoError, ErrorCode.NotImplemented])
@@ -137,20 +137,23 @@ class MachineTest(unittest.TestCase):
             self.assertEqual(err, ErrorCode.NoError)
 
         self.cpmIoController.SaveStateOn(3000)
-        self.memoryController.Write(0x00FE, 0xD3)
-        self.memoryController.Write(0x00FF, 0xFD)
+        self.memoryController.Write(0x00FE, 0xD3, None)
+        self.memoryController.Write(0x00FF, 0xFD, None)
         err = self.memoryController.Load(self.programsDir + 'TST8080.COM', 0x0100)
         self.assertEqual(err, ErrorCode.NoError)
         err = self.machine.SetOptions(r'{"rom":{"file":[{"offset":0,"size":1727}]},"ram":{"block":[{"offset":1727,"size":256}]},"isrFreq":0}')
         self.assertEqual(err, ErrorCode.NoError)
-        self.machine.SetIoController(self.cpmIoController)
-        self.machine.Run(0x0100)
+        err = self.machine.AttachIoController(self.cpmIoController)
+        self.assertEqual(err, ErrorCode.NoError)
+        err = self.machine.Run(0x0100)
+        self.assertEqual(err, ErrorCode.NoError)
 
         if runAsync == True:
             self.machine.WaitForCompletion()
 
         self.cpmIoController.SaveStateOn(-1)
-        self.machine.Run(0x00FE)
+        err = self.machine.Run(0x00FE)
+        self.assertEqual(err, ErrorCode.NoError)
 
         if runAsync == True:
             self.machine.WaitForCompletion()
@@ -180,10 +183,10 @@ class i8080Test(unittest.TestCase):
     def setUp(self):
         self.programsDir = MachineTestDeps.programsDir
         self.memoryController = MemoryController()
-        self.cpmIoController = CpmIoController(self.memoryController)
+        self.cpmIoController = CpmIoController()
         self.machine = Make8080Machine()
-        self.machine.SetIoController(self.cpmIoController)
-        self.machine.SetMemoryController(self.memoryController)
+        self.machine.AttachIoController(self.cpmIoController)
+        self.machine.AttachMemoryController(self.memoryController)
         err = self.machine.SetOptions(r'{"isrFreq":0.02}')
         self.assertEqual(err, ErrorCode.NoError)
         err = self.memoryController.Load(self.programsDir + 'exitTest.bin', 0x0000)
@@ -197,35 +200,30 @@ class i8080Test(unittest.TestCase):
         self.assertEqual(e, a['cpu'])
         ErrorCode.NoError
 
-    def test_8080Pre(self):
-        err = self.memoryController.Load(self.programsDir + '8080PRE.COM', 0x0100)
+    def RunTestSuite(self, suiteName, expectedState, expectedMsg):
+        err = self.memoryController.Load(self.programsDir + suiteName, 0x0100)
         self.assertEqual(err, ErrorCode.NoError)
-        self.machine.OnSave(lambda x: self.CheckMachineState(r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":0,"c":9,"d":3,"e":50,"h":1,"l":0,"s":86},"pc":2,"sp":1280}', x))
-        self.machine.Run(0x0100)
-        self.assertIn('8080 Preliminary tests complete', self.cpmIoController.Message())
+        err = self.machine.OnSave(lambda actualState: self.CheckMachineState(expectedState, actualState))
+        self.assertIn(err, [ErrorCode.NoError, ErrorCode.NotImplemented])
+        err = self.machine.Run(0x0100)
+        self.assertEqual(err, ErrorCode.NoError)
+
+        if suiteName == '8080EXM.COM':
+            self.assertNotIn(expectedMsg, self.cpmIoController.Message())
+        else:
+            self.assertIn(expectedMsg, self.cpmIoController.Message())
+
+    def test_8080Pre(self):
+        self.RunTestSuite('8080PRE.COM', r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":0,"c":9,"d":3,"e":50,"h":1,"l":0,"s":86},"pc":2,"sp":1280}', '8080 Preliminary tests complete')
 
     def test_Tst8080(self):
-        err = self.memoryController.Load(self.programsDir + 'TST8080.COM', 0x0100)
-        self.assertEqual(err, ErrorCode.NoError)
-        self.machine.OnSave(lambda x: self.CheckMachineState(r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":170,"b":170,"c":9,"d":170,"e":170,"h":170,"l":170,"s":86},"pc":2,"sp":1981}', x))
-        self.machine.Run(0x0100)
-        self.assertIn('CPU IS OPERATIONAL', self.cpmIoController.Message())
+        self.RunTestSuite('TST8080.COM', r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":170,"b":170,"c":9,"d":170,"e":170,"h":170,"l":170,"s":86},"pc":2,"sp":1981}', 'CPU IS OPERATIONAL')
 
-    # this will take a little while to complete
     def test_CpuTest(self):
-        err = self.memoryController.Load(self.programsDir + 'CPUTEST.COM', 0x0100)
-        self.assertEqual(err, ErrorCode.NoError)
-        self.machine.OnSave(lambda x: self.CheckMachineState(r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":0,"c":247,"d":4,"e":23,"h":0,"l":0,"s":70},"pc":2,"sp":12283}', x))
-        self.machine.Run(0x0100)
-        self.assertIn('CPU TESTS OK', self.cpmIoController.Message())
+        self.RunTestSuite('CPUTEST.COM', r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":0,"c":247,"d":4,"e":23,"h":0,"l":0,"s":70},"pc":2,"sp":12283}', 'CPU TESTS OK')
 
-    # this will take a long time to complete
     def test_8080Exm(self):
-        err = self.memoryController.Load(self.programsDir + '8080EXM.COM', 0x0100)
-        self.assertEqual(err, ErrorCode.NoError)
-        self.machine.OnSave(lambda x: self.CheckMachineState(r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":10,"c":9,"d":14,"e":30,"h":1,"l":109,"s":70},"pc":2,"sp":54137}', x))
-        self.machine.Run(0x0100)
-        self.assertNotIn('ERROR', self.cpmIoController.Message())
+        self.RunTestSuite('8080EXM.COM', r'{"uuid":"O+hPH516S3ClRdnzSRL8rQ==","registers":{"a":0,"b":10,"c":9,"d":14,"e":30,"h":1,"l":109,"s":70},"pc":2,"sp":54137}', 'ERROR')
 
 if __name__ == '__main__':
     unittest.main()